split big augeas_parser spec
